Skip workflow for project creation
When importing a full project data object - which can be large ~5mb - then running through the workflow engine will be inefficient and will chew data.

Skip workflow rules for project create calls and go straight to the resolver.
