+++ lib/api.js
@@ -18,7 +18,7 @@ const metricsRouter = require('./router/metrics');
 const decorators = require('./decorators');
 
 const profile = require('./middleware/profile');
-const resolveProjectVersionUpdates = require('./middleware/resolve-project-version-updates');
+const skipWorkflow = require('./middleware/skip-workflow');
 
 const hooks = {
   unique: require('./hooks/create/unique'),
@@ -43,7 +43,7 @@ module.exports = settings => {
 
   app.use(profile(settings));
 
-  app.use(resolveProjectVersionUpdates(settings));
+  app.use(skipWorkflow(settings));
 
   const flow = Taskflow({ db: settings.taskflowDB });
