+++ lib/hooks/resolve/index.js
@@ -8,9 +8,12 @@ module.exports = settings => {
     if (settings.noDownstream) {
       return Promise.resolve();
     }
-
     return messager(model.data)
-      .then(changelogModel => model.patch({ id: changelogModel.modelId }));
+      .then(changelogModel => {
+        const id = changelogModel.modelId;
+        const result = model.patch({ id });
+        return result || { id };
+      });
   };
 
 };
